refactor: kakao client 폴더 이동
