+++ src/main/java/com/moyorak/api/restaurant/dto/RestaurantSearchRequest.java
@@ -1,6 +1,6 @@
 package com.moyorak.api.restaurant.dto;
 
-import com.moyorak.api.restaurant.external.dto.KakaoSearchRequest;
+import com.moyorak.infra.client.kakao.dto.KakaoSearchRequest;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.*;
