+++ src/main/java/com/moyorak/api/restaurant/service/RestaurantService.java
@@ -2,10 +2,10 @@ package com.moyorak.api.restaurant.service;
 
 import com.moyorak.api.restaurant.dto.RestaurantResponse;
 import com.moyorak.api.restaurant.dto.RestaurantSearchRequest;
-import com.moyorak.api.restaurant.external.KakaoSearcher;
-import com.moyorak.api.restaurant.external.dto.KakaoPlace;
-import com.moyorak.api.restaurant.external.dto.KakoSearchResponse;
 import com.moyorak.global.domain.ListResponse;
+import com.moyorak.infra.client.kakao.KakaoSearcher;
+import com.moyorak.infra.client.kakao.dto.KakaoPlace;
+import com.moyorak.infra.client.kakao.dto.KakoSearchResponse;
 import java.util.Collections;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
