+++ src/main/java/com/moyorak/infra/client/kakao/dto/KakaoMeta.java
@@ -1,3 +1,3 @@
-package com.moyorak.api.restaurant.external.dto;
+package com.moyorak.infra.client.kakao.dto;
 
 public record KakaoMeta(int total_count, int pageable_count, boolean is_end) {}
