+++ src/main/java/com/moyorak/infra/client/kakao/dto/KakaoPlace.java
@@ -1,4 +1,4 @@
-package com.moyorak.api.restaurant.external.dto;
+package com.moyorak.infra.client.kakao.dto;
 
 public record KakaoPlace(
         String place_name,
