+++ src/main/java/com/moyorak/infra/client/kakao/dto/KakoSearchResponse.java
@@ -1,4 +1,4 @@
-package com.moyorak.api.restaurant.external.dto;
+package com.moyorak.infra.client.kakao.dto;
 
 import java.util.List;
